backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
alembic==1.15.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
email-validator==2.2.0
orjson==3.10.15
httpx==0.28.1
jinja2==3.1.5
pillow==11.1.0
srt==3.5.3
openai==1.14.2
edge-tts==6.1.13
mutagen==1.47.0